[Style] 함수 분리 및 패키지화
