+++ src/main/kotlin/christmas/SaleController.kt
@@ -1,0 +1,10 @@
+package christmas
+
+import christmas.view.InputView
+
+class SaleController {
+
+//입력한 요일을 판단하는 함수
+
+    val userInput = InputView()
+}
