+++ src/main/kotlin/christmas/domain/Order.kt
@@ -1,0 +1,5 @@
+package christmas.domain
+
+data class Order(val menuName: String, val quantity: Int) {
+//    init
+}
