+++ src/main/kotlin/christmas/util/Appetizer.kt
@@ -1,0 +1,7 @@
+package christmas.util
+
+enum class Appetizer(val appetizerName: String, val appetizerMoney: Int) {
+    MUSHROOM_SOUP("양송이수프", 6_000),
+    TAPAS("타파스", 5_500),
+    CAESAR_SALAD("시저샐러드", 8_000);
+}
