+++ src/main/kotlin/christmas/util/Beverage.kt
@@ -1,0 +1,7 @@
+package christmas.util
+
+enum class Beverage(val drinkName: String, val drinkPrice: Int) {
+    ZERO_COLA("제로콜라", 3_000),
+    RED_WINE("레드와인", 60_000),
+    CHAMPAGNE("샴페인", 25_000);
+}
