+++ src/main/kotlin/christmas/util/DayOfWeek.kt
@@ -1,0 +1,11 @@
+package christmas.util
+
+enum class DayOfWeek(val dayNumber: Int, val dayName: String) {
+    FRIDAY(1, "금요일"),
+    SATURDAY(2, "토요일"),
+    SUNDAY(3, "일요일"),
+    MONDAY(4, "월요일"),
+    TUESDAY(5, "화요일"),
+    WEDNESDAY(6, "수요일"),
+    THURSDAY(7, "목요일");
+}
