+++ src/main/kotlin/christmas/util/Dessert.kt
@@ -1,0 +1,6 @@
+package christmas.util
+
+enum class Dessert(val dessertName: String, val dessertPrice: Int) {
+    CHOCO_CAKE("초코케이크", 15_000),
+    ICE_CREAM("아이스크림", 5_000);
+}
