+++ src/main/kotlin/christmas/util/MainMenu.kt
@@ -1,0 +1,8 @@
+package christmas.util
+
+enum class MainMenu(val mainMenuName: String, val mainMenuMoney: Int) {
+    T_BONE_STEAK("티본스테이크", 55_000),
+    BBQ_RIB("바비큐립", 54_000),
+    SEAFOOD_PASTA("해산물파스타", 35_000),
+    CHRISTMAS_PASTA("크리스마스파스타", 25_000);
+}
