+++ src/main/kotlin/christmas/view/InputView.kt
@@ -1,4 +1,4 @@
-package christmas
+package christmas.view
 
 import camp.nextstep.edu.missionutils.Console
